backend/src/main/java/hw/spring/services/JwtTokenHandler.java
package hw.spring.services;

import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.stereotype.Component;

import javax.inject.Inject;
import java.time.ZonedDateTime;
import java.util.Date;
import java.util.Optional;

@Component
public final class JwtTokenHandler {

    private final String secret;
    @Inject UserDetailsService userService;

    public JwtTokenHandler(@Value("${hw.spring.jwt.secret}") String secret) {
        this.secret = secret;
    }

    Optional<UserDetails> parseUserFromToken(String token) {
        String username = Jwts.parser()
                .setSigningKey(secret)
                .parseClaimsJws(token)
                .getBody()
                .getSubject();
        return Optional.ofNullable(userService.loadUserByUsername(username));
    }

    public String createTokenForUser(UserDetails user) {
        final ZonedDateTime afterOneWeek = ZonedDateTime.now().plusWeeks(1);
        return Jwts.builder()
                .setSubject(user.getUsername())
                .signWith(SignatureAlgorithm.HS256, secret)
                .setExpiration(Date.from(afterOneWeek.toInstant()))
                .compact();
    }
}